refactor: arrange folder structure
