+++ src/lib/image-event-handler.ts
@@ -1,6 +1,6 @@
 import type { App } from 'obsidian';
+import type DoubleClickImageOpenerPlugin from '../main';
 import { ErrorHandler } from './error-handler';
-import type DoubleClickImageOpenerPlugin from './main';
 import { PathResolver } from './path-resolver';
 import { openWithDefaultApp } from './system-launcher';
