+++ src/lib/path-resolver.test.ts
