+++ src/lib/path-resolver.ts
